refactor: keep refusal wording with the operation, not the entry point
Ending a reservation early now has two entry points, and the second was
borrowing the first one's wording through pub(super). What a refusal means
depends on the operation, not on which button was pressed, so both the
end-now and the cancel wording move next to each other.
